requirements/production.txt
click
Flask==1.1.2
flask_apispec==0.11.0
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
flask_restful
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
marshmallow==3.10.0
psycopg2-binary==2.8.6
python-slugify==4.0.1
requests
simplejson
SQLAlchemy==1.3.23
uwsgi
Werkzeug==1.0.1